move get_view_direction_in_scene_coordinates()
